requirements.txt
PyJWT~=2.6.0
chalice~=1.27.3
boto3~=1.26.13
botocore~=1.29.13